refactoring-parser: part 4 - add struct logging - add request id logging - add rewrite filter to rewrite telegram id to user id - add timetable source for get in composer - add abstract class of message broker - change naming of notifications - refactoring kafka config - prep for new notifications about timetables

+++ src/main/kotlin/com/melowetty/hsepermhelper/service/impl/timetable/ExcelTimetableProcessChangesService.kt
@@ -10,6 +10,8 @@ class ExcelTimetableProcessChangesService(
     private val storage: ExcelTimetableStorage,
     private val timetableNotificationService: TimetableNotificationService,
 ) {
+    // TODO: сделать запуск задач по проверке изменений и отправке уведомлений
+
     fun processChangesByRunIds(prevRunId: String, currentRunId: String) {
         val oldTimetables = storage.getTimetablesInfoIdByRunId(prevRunId)
         val oldTimetablesIds = oldTimetables.map { it.id }.toSet()
@@ -44,7 +46,9 @@ class ExcelTimetableProcessChangesService(
 
         storage.showTimetables(currentTimetablesIds.toList())
 
-        storage.deleteTimetablesByRunId(prevRunId)
+        if (notChangedTimetables.size != currentTimetables.size) {
+            storage.deleteTimetablesByRunId(prevRunId)
+        }
     }
 
     private fun processTimetables(
